requirements.txt
asgiref
Django
django-recurrence
djangorestframework
Pillow
python-dateutil
pytz
six
sqlparse
tzdata
